Remove :rig_tests OTP app and move code into :rig
