+++ apps/rig/config/config.exs
@@ -71,3 +71,4 @@ import_config "rig_auth/config.exs"
 import_config "rig_inbound_gateway/config.exs"
 import_config "rig_metrics/config.exs"
 import_config "rig_outbound_gateway/config.exs"
+import_config "rig_tests/config.exs"
