--- apps/rig/lib/rig_auth/blacklist/sup.ex
@@ -1,26 +1,0 @@
-defmodule RigAuth.Blacklist.Sup do
-  @moduledoc """
-  Supervisor for the global Blacklist.
-
-  """
-
-  use Supervisor
-
-  alias RigAuth.Blacklist.PresenceHandler
-
-  def start_link(_) do
-    Supervisor.start_link(__MODULE__, :ok)
-  end
-
-  @impl Supervisor
-  def init(:ok) do
-    children = [
-      # The PresenceHandler also starts the Blacklist server. Since that's kind
-      # of hard to guess, this supervisor exists to hide this implementation
-      # detail to the parent supervisor.
-      worker(PresenceHandler, _args = [[pubsub_server: Rig.PubSub]])
-    ]
-
-    supervise(children, strategy: :one_for_one)
-  end
-end
