+++ apps/rig/lib/rig_inbound_gateway_web.ex
@@ -31,12 +31,6 @@ defmodule RigInboundGatewayWeb do
     end
   end
 
-  def channel do
-    quote do
-      use Phoenix.Channel
-    end
-  end
-
   @doc """
   When used, dispatch to the appropriate controller/view/etc.
   """
