+++ apps/rig/test/rig_tests/proxy/response_from/kafka_test.exs
@@ -1,32 +1,70 @@
-defmodule RigTests.Proxy.ResponseFrom.AsyncHttpTest do
+defmodule RigTests.Proxy.ResponseFrom.KafkaTest do
   @moduledoc """
-  If `response_from` is set to http_async, the response is taken from internal HTTP endpoint /v1/responses
+  With `response_from` set to Kafka, the response is taken from a Kafka topic.
+
+  In production, this may be used to hide asynchronous processing by one or more
+  backend services with a synchronous interface.
 
   Note that `test_with_server` sets up an HTTP server mock, which is then configured
   using the `route` macro.
   """
-  # cause FakeServer opens a port:
+  use Rig.Config, [
+    :brokers,
+    :consumer_topics,
+    :ssl_enabled?,
+    :ssl_ca_certfile,
+    :ssl_certfile,
+    :ssl_keyfile,
+    :ssl_keyfile_pass,
+    :sasl,
+    :response_topic
+  ]
+
   use ExUnit.Case, async: false
-
   import FakeServer
-  import Plug.Conn, only: [put_req_header: 3]
-  import Phoenix.ConnTest, only: [post: 3, build_conn: 0]
+  alias FakeServer.HTTP.Response
 
-  alias FakeServer.Response
+  alias Rig.KafkaConfig, as: RigKafkaConfig
+  alias RigInboundGateway.ApiProxyInjection
+  alias RigKafka
 
-  @endpoint RigApi.Endpoint
   @api_port Confex.fetch_env!(:rig, RigApi.Endpoint)[:http][:port]
   @proxy_port Confex.fetch_env!(:rig, RigInboundGatewayWeb.Endpoint)[:http][:port]
 
-  test_with_server "Given response_from is set to http_async, the http response is taken from the internal HTTP endpoint." do
-    test_name = "proxy-http-response-from-http-internal"
+  defp kafka_config, do: RigKafkaConfig.parse(config())
+
+  setup_all do
+    ApiProxyInjection.set()
+
+    on_exit(fn ->
+      ApiProxyInjection.restore()
+    end)
+  end
+
+  setup do
+    kafka_config = kafka_config()
+    {:ok, kafka_client} = RigKafka.start(kafka_config)
+
+    on_exit(fn ->
+      :ok = RigKafka.Client.stop_supervised(kafka_client)
+    end)
+
+    :ok
+  end
+
+  @tag :kafka
+  test_with_server "Given response_from is set to Kafka, the http response is taken from the Kafka response topic instead of forwarding the backend's original response." do
+    test_name = "proxy-http-response-from-kafka"
 
     api_id = "mock-#{test_name}-api"
     endpoint_id = "mock-#{test_name}-endpoint"
+    %{response_topic: kafka_topic} = config()
     endpoint_path = "/#{endpoint_id}"
-    sync_response = %{"this response" => "the client never sees this response"}
+    sync_response = %{"message" => "the client never sees this response"}
     async_response = %{"message" => "this is the async response that reaches the client instead"}
 
+    # The following service fake also shows how a real service should
+    # wrap its response in a CloudEvent:
     route(endpoint_path, fn %{query: %{"correlation" => correlation_id}} ->
       event =
         Jason.encode!(%{
@@ -38,10 +76,8 @@ defmodule RigTests.Proxy.ResponseFrom.AsyncHttpTest do
           data: async_response
         })
 
-      build_conn()
-      |> put_req_header("content-type", "application/json;charset=utf-8")
-      |> post("/v1/responses", event)
-
+      kafka_config = kafka_config()
+      assert :ok == RigKafka.produce(kafka_config, kafka_topic, "", "response", event)
       Response.ok!(sync_response, %{"content-type" => "application/json"})
     end)
 
@@ -59,9 +95,10 @@ defmodule RigTests.Proxy.ResponseFrom.AsyncHttpTest do
               %{
                 id: endpoint_id,
                 type: "http",
+                secured: false,
                 method: "GET",
                 path: endpoint_path,
-                response_from: "http_async"
+                response_from: "kafka"
               }
             ]
           }
@@ -77,6 +114,7 @@ defmodule RigTests.Proxy.ResponseFrom.AsyncHttpTest do
 
     # The client calls the proxy endpoint:
     request_url = rig_proxy_url <> endpoint_path
+
     %HTTPoison.Response{status_code: res_status, body: res_body} = HTTPoison.get!(request_url)
 
     # Now we can assert that...
@@ -86,7 +124,7 @@ defmodule RigTests.Proxy.ResponseFrom.AsyncHttpTest do
     assert res_status == 200
     # ...the client never saw the http response:
     assert Jason.decode!(res_body) != sync_response
-    # ...but the client got the response sent to the HTTP internal endpoint:
+    # ...but the client got the response sent to the Kafka topic:
     assert Jason.decode!(res_body)["data"] == async_response
   end
 end
