--- apps/rig_tests/README.md
@@ -1,3 +1,0 @@
-# RIG Tests
-
-Tests that span multiple umbrella apps.
