--- apps/rig_tests/mix.exs
@@ -1,43 +1,0 @@
-defmodule RigTests.MixProject do
-  @moduledoc false
-  use Mix.Project
-
-  def project do
-    %{rig: rig_version, elixir: elixir_version} = versions()
-
-    [
-      app: :rig_tests,
-      version: rig_version,
-      build_path: "../../_build",
-      config_path: "../../config/config.exs",
-      deps_path: "../../deps",
-      lockfile: "../../mix.lock",
-      elixir: elixir_version,
-      elixirc_paths: elixirc_paths(Mix.env()),
-      start_permanent: Mix.env() == :prod,
-      deps: deps(),
-      test_coverage: [tool: ExCoveralls]
-    ]
-  end
-
-  def application do
-    [
-      extra_applications: [:logger]
-    ]
-  end
-
-  defp versions do
-    {map, []} = Code.eval_file("version", "../..")
-    map
-  end
-
-  defp elixirc_paths(:test), do: ["lib", "test/support"]
-  defp elixirc_paths(_), do: ["lib"]
-
-  defp deps do
-    [
-      {:rig, in_umbrella: true},
-      {:fake_server, "~> 2.0", only: :test}
-    ]
-  end
-end
